requirements.txt
fastapi
pydantic
crewai
google-generativeai
python-dotenv
uvicorn
